+++ source/ChanSort.Loader.Panasonic/PanasonicPlugin.cs
@@ -12,7 +12,20 @@ namespace ChanSort.Loader.Panasonic
 
     public SerializerBase CreateSerializer(string inputFile)
     {
-      if (Path.GetExtension(inputFile).ToLowerInvariant() == ".xml")
+      // check for files in the 2022 /mnt/vendor/tvdata/database/channel/ directory structure file format with tv.db and idtvChannel.bin
+      var name = Path.GetFileName(inputFile).ToLowerInvariant();
+      var baseDir = Path.GetDirectoryName(inputFile);
+      if (name == "hotel.bin")
+        baseDir = Path.Combine(baseDir, "mnt", "vendor", "tvdata", "database");
+      if (name == "idtvChannel.bin")
+        baseDir = Path.GetDirectoryName(baseDir);
+      var tvDb = Path.Combine(baseDir, "tv.db");
+      if (File.Exists(tvDb) && File.Exists(Path.Combine(baseDir, "channel", "idtvChannel.bin")))
+        return new IdtvChannelSerializer(tvDb);
+
+      // Android based models use an .xml format. Unfortunately that format is utter garbage and not really useful
+      var ext = Path.GetExtension(inputFile).ToLowerInvariant();
+      if (ext == ".xml")
       {
         var data = File.ReadAllBytes(inputFile);
         var header = Encoding.ASCII.GetBytes("<ChannelList>\n<ChannelInfo IsModified=");
@@ -25,6 +38,7 @@ namespace ChanSort.Loader.Panasonic
         return new XmlSerializer(inputFile);
       }
 
+      // old svl.db / svl.bin formats
       return new Serializer(inputFile);
     }
   }